Fixing Package-Create for debian on ARM and preinst script

+++ package/create_package.sh
@@ -11,6 +11,9 @@ cd - &>/dev/null
 PACKAGE="network_discover"
 VERSION=$( grep "^version" ../Cargo.toml | cut -d'"' -f2 )
 ARCH=$( uname -m )
+if [ "${ARCH::3}" == "arm" ]; then
+	ARCH="armhf"
+fi
 ARCHIVE="${PACKAGE}-${VERSION}_${ARCH}.tar.xz"
 DEBIAN="${PACKAGE}-${VERSION}_${ARCH}.deb"
 
@@ -56,9 +59,9 @@ $( cd pkg; find . -type f | xargs md5sum | sed 's/.\///' > ../deb/md5sums )
 cat << EOF > deb/control
 Package: ${PACKAGE}
 Version: ${VERSION}-1
-Architecture: amd64
+Architecture: ${ARCH}
 Maintainer: Lukas LukyLuke Zurschmiede <${PACKAGE}@ranta.ch>
-Depends: nmap ( >= 7.91 )
+Depends: nmap ( >= 7.40 )
 Section: net
 Priority: optional
 Homepage: https://github.com/ITSGmbH/NetworkDiscover
@@ -107,7 +110,9 @@ EOF
 cat << EOF > deb/preinst
 #!/bin/sh
 set -e
-systemctl stop ${PACKAGE}
+if [ -f /lib/systemd/system/${PACKAGE}.service ]; then
+  systemctl stop ${PACKAGE}
+fi
 exit 0
 EOF
